Repo's and Services added. Next up - build controllers and test endpoints...

+++ src/main/java/maddogsoftworks/cocktail_codex/repository/CustomerRepository.java
@@ -3,5 +3,7 @@ package maddogsoftworks.cocktail_codex.repository;
 import maddogsoftworks.cocktail_codex.model.Customer;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CustomerRepository extends JpaRepository<Customer, Integer> {
+import java.util.List;
+
+public interface CustomerRepository extends JpaRepository<Customer, Long> {
 }
